Removed flint version check in `finitefield.py`

+++ sympy/ntheory/factor_.py
@@ -963,7 +963,7 @@ def _check_termination(factors, n, limitp1, use_trial, use_rho, use_pm1,
         for b, e in facs.items():
             if verbose:
                 print(factor_msg % (b, e))
-            factors[b] = exp*e
+            factors[b] = int(exp*e)
         if verbose:
             print(complete_msg)
         return True
